setup-analysis/pyproject.toml
[project]
name = "setup-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mne>=1.7.1",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
    "matplotlib>=3.9.0",
    "ipykernel>=6.29.4",
    "statsmodels>=0.14.2",
    "scipy>=1.14.0",
    "tqdm>=4.66.4",
    "tabulate>=0.9.0",
    "mne-icalabel>=0.6.0",
    "screeninfo>=0.8.1",
    "icecream>=2.1.3",
    "pillow>=10.3.0",
    "openpyxl>=3.1.4",
    "mne-qt-browser>=0.6.3",
    "pyqt6>=6.7.0",
    "ipywidgets>=8.1.3",
    "nibabel>=5.2.1",
    "polars>=1.8.1",
    "scikit-learn>=1.5.2",
    "python-box>=7.2.0",
    "plotly>=5.24.1",
    "nbformat>=5.10.4",
    "dash>=2.18.1",
    "dash-ag-grid>=31.2.0",
    "mplcursors>=0.6",
    "ipympl>=0.9.4",
    "pyprep>=0.4.3",
    "pympler>=1.1",
    "rsatoolbox>=0.2.0",
    "mne-rsa>=0.9.1",
    "neurora>=1.1.6.12",
    "pylustrator>=1.3.0",
    "hmp>=0.4.1",
    "tensorpac>=0.6.5",
    "directory-tree>=1.0.0",
    "autoreject>=0.4.3",
    "imgkit>=1.2.3",
    "pyperclip>=1.9.0",
    "pendulum>=3.0.0",
    "loguru>=0.7.2",
    "rich>=13.7.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "dataframe-image>=0.2.6",
    "pingouin>=0.5.5",
    "transformers>=4.48.3",
    "python-pptx>=1.0.2",
    "lazypredict>=0.2.13",
    "bambi>=0.15.0",
    "arviz>=0.20.0",
]
